LayoutTests/media/media-vp8-webm-initialcurrentTime-expected.txt


EVENT(loadedmetadata)
EXPECTED (0 == '0') OK
